### 2024-01-26 #6287 @kyleniemeyer
Hello @KskAdch, thanks for your submission to JOSS. We'll use this pre-review issue to assign an editor and find reviewers.

Unfortunately, we don't currently have an editor in this area available to handle your submission, so I'll put this on our waitlist until someone becomes available. In the meantime, any reviewer recommendations you can make would be welcome.
