### 2024-04-21 #6287 @phjmsycc
Hi, @olexandr-konovalov. I'm a co-researcher on this project. Since @KskAdch graduated and left our laboratory this spring, I will primarily handle the revision of this work. As far as I know, you were assigned as the editor on February 9th, and now, over two months have passed. Could you please provide an update on the status of the review?
