### 2024-05-22 #6287 @olexandr-konovalov
Hello @Datseris, would you be able to review this submission for JOSS, please?

If not, could you possibly recommend someone whom we may approach regarding this?
