### 2024-05-25 #6287 @olexandr-konovalov
Hi @Chronum94 and @santoshkumarradha - would one or both of you be able to review this submission for JOSS, please?

The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. JOSS reviews involve downloading and installing the software, and inspecting the repository and submitted paper for key elements (see https://joss.readthedocs.io/en/latest/review_criteria.html). Editors and reviewers post comments on the Review issue, and authors respond to the comments and improve their submission until acceptance (or withdrawal, if they feel unable to satisfy the review). You can find more details in the reviewer guides at https://joss.readthedocs.io/.

If not, could you possibly recommend someone whom we may approach regarding this? Thank you!
