### 2024-06-04 #6287 @olexandr-konovalov
Hi @vavrines and @zafeirakopoulos - would one or both of you be able to review this submission for JOSS, please? If not, could you possibly suggest someone whom we may approach regarding this?
