### 2024-06-19 #6287 @olexandr-konovalov
Hi @davibarreira and @brainandforce - would one or both of you be able to review this submission for JOSS, please? Please see details about our review process a few comments above. If not, could you possibly suggest someone whom we may approach regarding this?
