### 2024-06-19 #6287 @davibarreira
@olexandr-konovalov , I'd love to help, but after taking a look at the package, this seems way out of my expertise. Although I do know some Topology, it is the first time I hear of topological numbers. I also don't know anybody with knowledge in this subject.
