### 2024-06-20 #6287 @brainandforce
I am *brand new* to the field of topological data analysis and I would not be suited to verify anything about the software outside of its overall organization.
