### 2024-06-30 #6287 @michiexile
> Hi @michiexile - long time since we been in touch! I am looking for reviewers of this JOSS paper about topological data analysis in Julia. Would you be interested, and/or possibly suggest someone else to review it? Please see details about our review process a few comments above.

Sorry for not responding earlier. I'm happy to review.
