### 2024-06-30 #6944 @editorialbot
[REVIEW]: TopologicalNumbers.jl: A Julia package for topological number computation
**Submitting author:** @KskAdch (<a href="http://orcid.org/0009-0004-0195-7952">Keisuke Adachi</a>)
**Repository:** https://github.com/KskAdch/TopologicalNumbers.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.7.6
