### 2024-07-02 #6944 @michiexile
In my opinion, the two missing check-marks in my checklist are not blocking acceptance of the paper. The authors do describe what the software is doing, well enough for anyone who knows the domain they are writing for, and they do describe the lack of Julia packages for doing the same computations (however, no discussion of packages for other platforms).
