### 2024-07-23 #6944 @olexandr-konovalov
Thank you @michiexile! Perhaps let's wait till @lf28 completes his review, to decide if this is acceptable, or some improvements would still be possible and make sense.
