### 2024-10-04 #6944 @olexandr-konovalov
Thank you for your reviews @michiexile and @lf28 - also apologise to you and @KskAdch for the delay from my side in responding to you. I am glad that you agree in your opinions, and also that except the two remaining checkboxes, everything else looks good. But to accept the paper, we need also to be able to add those two ticks for the summary and the state of the field as well. As @michiexile says

>  The authors do describe what the software is doing, well enough for anyone who knows the domain they are writing for, and they do describe the lack of Julia packages for doing the same computations (however, no discussion of packages for other platforms).

but we indeed need a "clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience" and also "how this software compares to other commonly-used packages".

Can perhaps something still be done here, @KskAdch?
