+++ .github/workflows/CI.yml
@@ -19,8 +19,9 @@ jobs:
       matrix:
         version:
           - "1.6"
-          - "1.9"
-          - "1.10"
+          # - "1.9"
+          # - "1.10"
+          - "1.11"
           # - "nightly"
         os:
           - ubuntu-latest
