### 2024-11-28 b19065b
update acknowledgement and bib.
