+++ paper/paper.md
@@ -113,6 +113,4 @@ and we believe that it will be useful for researchers in the field of solid-stat
 
 # Acknowledgements
 The authors are grateful to Takahiro Fukui for fruitful discussions. 
-M.K. was supported by JST, 
-the establishment of university fellowships towards the creation of science technology innovation, 
-Grant No. JPMJFS2107.
+M. K. was supported by JST, the establishment of university fellowships towards the creation of science technology innovation (Grant No. JPMJFS2107), and by JST SPRING (Grant No. JPMJSP2109).
