+++ paper/paper.bib
@@ -1,162 +1,188 @@
-@article{Fukui2005Chern,
-  title = {Chern {{Numbers}} in {{Discretized Brillouin Zone}}: {{Efficient Method}} of {{Computing}} ({{Spin}}) {{Hall Conductances}}},
-  shorttitle = {Chern {{Numbers}} in {{Discretized Brillouin Zone}}},
-  author = {Fukui, Takahiro and Hatsugai, Yasuhiro and Suzuki, Hiroshi},
-  year = {2005},
-  journal = {J. Phys. Soc. Jpn.},
-  volume = {74},
-  number = {6},
-  pages = {1674--1677},
-  publisher = {{The Physical Society of Japan}},
-  issn = {0031-9015},
-  doi = {10.1143/JPSJ.74.1674}
-}
-
-@misc{Shiozaki2023discrete,
-  title = {A Discrete Formulation of the {{Kane-Mele}} ${\mathbb Z}_2$ Invariant},
-  author = {Shiozaki, Ken},
-  year = {2023},
-  number = {arXiv:2305.05615},
-  eprint = {2305.05615},
-  primaryclass = {cond-mat},
-  publisher = {{arXiv}},
-  doi = {10.48550/arXiv.2305.05615},
-  urldate = {2023-07-29},
-  archiveprefix = {arxiv}
-}
-
 @article{Du2017Emergence,
-  title = {Emergence of Topological Nodal Lines and Type-{{II Weyl}} Nodes in the Strong Spin-Orbit Coupling System InNbX2 (X = S,Se)},
-  author = {Du, Yongping and Bo, Xiangyan and Wang, Di and Kan, Er-jun and Duan, Chun-Gang and Savrasov, Sergey Y. and Wan, Xiangang},
-  year = {2017},
-  journal = {Phys. Rev. B},
-  volume = {96},
-  number = {23},
-  pages = {235152},
+  title     = {Emergence of Topological Nodal Lines and Type-{{II Weyl}} Nodes in the Strong Spin-Orbit Coupling System InNbX2 (X = S,Se)},
+  author    = {Du, Yongping and Bo, Xiangyan and Wang, Di and Kan, Er-jun and Duan, Chun-Gang and Savrasov, Sergey Y. and Wan, Xiangang},
+  year      = {2017},
+  journal   = {Phys. Rev. B},
   publisher = {{American Physical Society}},
-  doi = {10.1103/PhysRevB.96.235152},
-  urldate = {2024-01-23}
+  doi       = {10.1103/PhysRevB.96.235152},
+  number    = {23},
+  pages     = {235152},
+  urldate   = {2024-01-23},
+  volume    = {96},
 }
 
-@article{Hirayama2015Weyl,
-  title = {Weyl {{Node}} and {{Spin Texture}} in {{Trigonal Tellurium}} and {{Selenium}}},
-  author = {Hirayama, Motoaki and Okugawa, Ryo and Ishibashi, Shoji and Murakami, Shuichi and Miyake, Takashi},
-  year = {2015},
-  journal = {Phys. Rev. Lett.},
-  volume = {114},
-  number = {20},
-  pages = {206401},
-  publisher = {{American Physical Society}},
-  doi = {10.1103/PhysRevLett.114.206401},
-  urldate = {2024-01-23}
-}
-
-@article{Hirayama2018Topological,
-  title = {Topological {{Semimetals Studied}} by {{Ab Initio Calculations}}},
-  author = {Hirayama, Motoaki and Okugawa, Ryo and Murakami, Shuichi},
-  year = {2018},
-  journal = {J. Phys. Soc. Jpn.},
-  volume = {87},
-  number = {4},
-  pages = {041002},
[…]
