### 2024-12-02 841b7c3
Update CI configuration and README for Julia version compatibility
