+++ .github/workflows/CI.yml
@@ -18,9 +18,9 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - "1.6"
+          # - "1.6"
           # - "1.9"
-          # - "1.10"
+          - "1.10"
           - "1.11"
           # - "nightly"
         os:
