### 2024-12-02 991fcb7
Update compatibility versions in Project.toml for Julia 1.10
