### 2024-12-03 c171a6a
update paper draft
