### 2024-12-04 6e54d70
update paper draft
