+++ paper/paper.md
@@ -112,13 +112,13 @@ Other topological numbers similarly become important physical quantities that ch
 例えば、電気伝導度は物質が電気をどの程度通すかを示す重要な物理量であり、通常は、物質に弱い静電場をかけたとき物質中の大域的空間（バルク）において電子が遷移可能であるような量子状態が存在すれば有限の値を持ち、電気が流れる金属となります。
 逆に存在しなければ低い値を持ち、電気を流さない絶縁体となります。
 <!-- しかし、2005年に理論的に提唱されたトポロジカル絶縁体は、バルクでは絶縁体でありながら、物質表面には電子の伝導状態を持つ新しい物質状態を示します[Ref.]。 -->
-しかし、1980年以降、今日に至るまで大きな関心を集めている量子ホール絶縁体やトポロジカル絶縁体・超伝導体などを含むトポロジカル電子系は、バルクでは絶縁体でありながら、物質表面には電子の伝導状態を持つ新しい物質状態を示します[Ref.]。
+しかし、1980年以降、今日に至るまで大きな関心を集めている量子ホール絶縁体やトポロジカル絶縁体などを含むトポロジカル電子系は、バルクでは絶縁体でありながら、物質表面には電子の伝導状態を持つ新しい物質状態を示します[Ref.]。
 <!-- このような新奇特性からトポロジカル絶縁体は大きな関心を集めており、今日に至るまでトポロジカル絶縁体候補物質の探索や特性評価が盛んに行われています[@Hasan2010Colloquium;@Qi2011Topological]。 -->
-このような新奇特性から、トポロジカル電子系の候補物質の探索や特性評価はこれまで盛んに行われて来ています[@Hasan2010Colloquium;@Qi2011Topological]。
+このような新奇特性から、トポロジカル電子系の候補物質の探索や特性評価がこれまで盛んに行われて来ています[@Hasan2010Colloquium;@Qi2011Topological]。
 
 
 
-表面の伝導状態の特徴は、物質のハミルトニアンから決定される固有量子状態が波数空間上でどのようなトポロジーを持つかによって決定されており、第一チャーン数、第二チャーン数、$\mathbb{Z}_2$不変量などのトポロジカル数がそれらを特徴付けるために用いられます。
+表面の伝導状態の特徴は、物質のハミルトニアンから決定される固有量子状態が波数空間上でどのようなトポロジーを持つかによって決定されており、第一チャーン数、第二チャーン数、$\mathbb{Z}_2$不変量などのトポロジカル数がそれらを特徴付けるために用いられます[Ref.]。
 典型的な例として、2次元物質に対して弱い電場をかけた際、かけた電場と直交方向に量子化された有限の電気伝導度（ホール伝導度）が現れる現象を量子ホール効果と言います[Ref.]。
 ホール伝導度$\sigma_{xy}$は整数値を取る第一チャーン数$\nu$によって特徴づけられ、$\sigma_{xy}=\frac{e^{2}}{h}\nu$と与えられることが知られています。
 ここで、$e$は電気素量、$h$はプランク定数です。
@@ -159,7 +159,7 @@ One is the method of second Chern number calculation in four-dimensional systems
 $\mathbb{Z}_2$ invariants can also be calculated in two-dimensional systems with time-reversal symmetry [@Fukui2007Quantum;@Shiozaki2023discrete]. 
 The FHS method is also applied to find Weyl points and Weyl nodes in three-dimensional systems [@Hirayama2018Topological;@Yang2011Quantum;@Hirayama2015Weyl;@Du2017Emergence].
 
[…]
