+++ paper/paper.md
@@ -36,50 +36,7 @@ bibliography: paper.bib
 
 
 # Summary
-<!--
-トポロジカル絶縁体はXXXX年に発見され盛んに研究されてきたが、絶縁体を含むトポロジカル物質がここ数年で大きな関心を集めている[@Hasan2010Colloquium;@Qi2011Topological]。
--->
-<!-- Topological insulators have been of considerable interest in the last decades [@Hasan2010Colloquium;@Qi2011Topological]. 
-These materials show new states of matter that are insulating in the bulk but have conducting states on their surfaces.  -->
-<!--
-ある対称性の下では、この表面上の伝導度はChern数や$\mathbb{Z}_{2}$数などによって計算できる。
-このChern数や$\mathbb{Z}_{2}$数は整数であり、まさしくトポロジカル数である。
-したがってトポロジカル数を計算することは、物質の表面の伝導度などの特徴的な物理量を計算することに匹敵する。
--->
-<!-- The conducting states on the surface are protected by the topology of the bulk band structure, and topological numbers, 
-such as first Chern number, second Chern number, $\mathbb{Z}_2$ number, etc., are used to characterize them. 
-As a typical example, the quantum Hall effect has the quantized Hall conductivity, which can be calculated by the first Chern number.  -->
-<!--
-実際に量子ホール効果が起こっているときのホール伝導度$\sigma$は、電子の素電化を$e$、プランク定数を$h$としたとき$\sigma=\tfrac{e^{2}}{h}\nu$と表せる。
-このときの$\nu$がトポロジカル数の一つである第一Chern数である。
--->
-<!-- Other topological numbers similarly become important physical quantities that characterize the system, 
-depending on the system dimension and symmetry classes [@Ryu2010Topological]. -->
-<!--
-さらに、これらのトポロジカル数は高度なエレクトロニクス、スピントロニクス、量子コンピューティングに応用できる可能性のあるトポロジカル絶縁体・超伝導体などのエキゾチックな物質を理解するために不可欠である。
--->
 
-<!-- To obtain the topological numbers, we often need numerical calculations, 
-and it may require an enormous amount of computation before convergence is achieved. 
-Therefore, creating tools to easily compute these numbers will lead to advances in research for the topological phase of matters. 
-So far, several methods have been reported that suggest that some topological numbers can be computed efficiently [@Fukui2005Chern;@Fukui2007Quantum;@Mochol-Grzelak2018Efficient;@Shiozaki2023discrete]. 
-However, since each method is specialized for a specific dimension or symmetry class, 
-it is necessary to implement the algorithm for each problem, respectively. 
-Our project, `TopologicalNumbers.jl`, aims to provide a package that can easily and efficiently compute topological numbers comprehensively in various dimensions and symmetry classes. -->
-<!--
-具体的には実空間ハミルトニアンをフーリエ変換した、波数空間におけるハミルトニアンと、計算したいトポロジカル数を指定する。
-この二つを指定するだけで次元や対称性クラスに依存することなく、包括的にトポロジカル数を計算することができる。
-また、莫大な時間を要する計算を上記の革新的な方法を用いて短縮することができる。
--->
-
-
-
-<!-- TopologicalNumbers.jl is an open-source Julia package designed to calculate topological invariants—mathematical quantities that characterize the global properties of materials—in condensed matter physics and related fields. 
[…]
