### 2024-12-04 f5dda4e
update paper draft
