+++ paper/paper.md
@@ -41,7 +41,7 @@ bibliography: paper.bib
 These invariants, such as the Chern number and $\mathbb{Z}_2$ invariants, are crucial for understanding exotic materials like topological insulators and superconductors which have potential applications in advanced electronics, spintronics, and quantum computing. 
 Our package provides an accessible and efficient toolset for researchers and educators to compute these invariants across various dimensions and symmetry classes, facilitating the exploration and discovery of new topological phases of matter.
 
-TopologicalNumbers.jlは、凝縮系物理学における物質の性質を特徴付ける数学的量であるトポロジカル不変量を計算するための、オープンソースのJuliaパッケージです。
+`TopologicalNumbers.jl`は、凝縮系物理学における物質の性質を特徴付ける数学的量であるトポロジカル不変量を計算するための、オープンソースのJuliaパッケージです。
 これらの不変量、例えばチャーン数や$\mathbb{Z}_2$不変量は、トポロジカル絶縁体や超伝導体のようなエキゾチックな物質を理解する上で重要であり、これらは先進的なエレクトロニクスやスピントロニクス、量子コンピューティングへの応用が期待されています[Ref.]。
 本パッケージは、研究者や教育者がこれらの不変量を様々な次元や対称性クラスで計算できる、使いやすく効率的なツールセットを提供し、新たなトポロジカル相の探索と発見を促進します。
 
@@ -73,7 +73,7 @@ Our project, `TopologicalNumbers.jl`, aims to provide a package that can easily 
 したがって、これらの数を容易に計算できるツールを作成することは、物質のトポロジカル相の研究の進展につながります。
 これまで、いくつかのトポロジカル数を効率的に計算できることを示唆する手法が報告されています[@Fukui2005Chern;@Fukui2007Quantum;@Mochol-Grzelak2018Efficient;@Shiozaki2023discrete]。
 しかし、各手法は特定の次元や対称性クラスに特化しているため、それぞれの問題に対してアルゴリズムを実装する必要があります。
-我々のプロジェクトであるTopologicalNumbers.jlは、様々な次元や対称性クラスにおいて、トポロジカル数を容易かつ効率的に包括的に計算できるパッケージを提供することを目指しています。
+我々のプロジェクトである`TopologicalNumbers.jl`は、様々な次元や対称性クラスにおいて、トポロジカル数を容易かつ効率的に包括的に計算できるパッケージを提供することを目指しています。
 
 
 This package currently includes various methods for calculating topological numbers. 
@@ -101,8 +101,8 @@ For instance, `Z2Pack` [@Gresch2017Z2Pack] is a Python-based tool widely used fo
 
 現在、これらすべての計算法を実装した包括的なJuliaパッケージは存在しません。
 一方、他のプラットフォームでは、本パッケージと異なるワニエ電荷中心[Ref.]またはWilson loop[Ref.]と呼ばれる量を用いた手法で実装されたいくつかのソフトウェアパッケージが利用可能です。
-例えば、Z2Pack [@Gresch2017Z2Pack]は、$\mathbb{Z}_2$不変量や第一チャーン数の計算に広く用いられるPythonベースのツールです。
-また、WannierTools [@Wu2018WannierTools]は、ワニエ関数を用いたトポロジカル物質の解析に強力な機能を提供しますが、Fortranで実装されており、一部のユーザーにとっては習得のハードルが高い可能性があります。
+例えば、`Z2Pack` [@Gresch2017Z2Pack]は、$\mathbb{Z}_2$不変量や第一チャーン数の計算に広く用いられるPythonベースのツールです。
[…]
