+++ paper/paper.bib
@@ -8,7 +8,7 @@
   number    = {23},
   pages     = {235152},
   urldate   = {2024-01-23},
-  volume    = {96},
+  volume    = {96}
 }
 
 @article{Fukui2005Chern,
@@ -22,7 +22,7 @@
   number     = {6},
   pages      = {1674--1677},
   shorttitle = {Chern {{Numbers}} in {{Discretized Brillouin Zone}}},
-  volume     = {74},
+  volume     = {74}
 }
 
 @article{Fukui2007Quantum,
@@ -36,7 +36,7 @@
   number     = {5},
   pages      = {053702},
   shorttitle = {Quantum {{Spin Hall Effect}} in {{Three Dimensional Materials}}},
-  volume     = {76},
+  volume     = {76}
 }
 
 @article{Gresch2017Z2Pack,
@@ -52,7 +52,7 @@
   pages      = {075146},
   shorttitle = {{{Z2Pack}}},
   urldate    = {2024-11-28},
-  volume     = {95},
+  volume     = {95}
 }
 
 @article{Hasan2010Colloquium,
@@ -66,7 +66,7 @@
   pages      = {3045--3067},
   shorttitle = {Colloquium},
   urldate    = {2024-01-22},
-  volume     = {82},
+  volume     = {82}
 }
 
 @article{Hirayama2015Weyl,
@@ -79,7 +79,7 @@
   number    = {20},
   pages     = {206401},
   urldate   = {2024-01-23},
-  volume    = {114},
+  volume    = {114}
 }
 
 @article{Hirayama2018Topological,
@@ -93,7 +93,7 @@
   number    = {4},
   pages     = {041002},
   urldate   = {2024-01-23},
-  volume    = {87},
+  volume    = {87}
 }
 
 @article{Mochol-Grzelak2018Efficient,
@@ -107,7 +107,7 @@
   langid    = {english},
   number    = {1},
   pages     = {014009},
-  volume    = {4},
+  volume    = {4}
 }
 
 @article{Qi2011Topological,
@@ -120,7 +120,7 @@
   number    = {4},
   pages     = {1057--1110},
   urldate   = {2024-01-22},
-  volume    = {83},
+  volume    = {83}
 }
 
 @article{Ryu2010Topological,
@@ -134,7 +134,7 @@
   pages      = {065010},
   shorttitle = {Topological Insulators and Superconductors},
   urldate    = {2024-01-23},
-  volume     = {12},
+  volume     = {12}
 }
 
 @misc{Shiozaki2023discrete,
@@ -147,7 +147,7 @@
   eprint        = {2305.05615},
   number        = {arXiv:2305.05615},
   primaryclass  = {cond-mat},
-  urldate       = {2023-07-29},
+  urldate       = {2023-07-29}
 }
 
 @article{Soluyanov2011Computing,
@@ -157,7 +157,7 @@
   journal = {Phys. Rev. B},
   doi     = {10.1103/PhysRevB.83.235401},
   number  = {23},
-  volume  = {83},
+  volume  = {83}
 }
 
 @article{Wimmer2012Algorithm,
@@ -170,7 +170,7 @@
   number     = {4},
   pages      = {30:1--30:17},
   shorttitle = {Algorithm 923},
-  volume     = {38},
+  volume     = {38}
 }
 
 @article{Yang2011Quantum,
@@ -184,5 +184,70 @@
   pages      = {075129},
   shorttitle = {Quantum {{Hall}} Effects in a {{Weyl}} Semimetal},
   urldate    = {2024-01-23},
-  volume     = {84},
+  volume     = {84}
+}
+
+@article{Yu2011Equivalent,
[…]
