+++ paper/paper.md
@@ -42,7 +42,7 @@ These invariants, such as the Chern number and $\mathbb{Z}_2$ invariants, are cr
 Our package provides an accessible and efficient toolset for researchers and educators to compute these invariants across various dimensions and symmetry classes, facilitating the exploration and discovery of new topological phases of matter.
 
 `TopologicalNumbers.jl`は、凝縮系物理学における物質の性質を特徴付ける数学的量であるトポロジカル不変量を計算するための、オープンソースのJuliaパッケージです。
-これらの不変量、例えばチャーン数や$\mathbb{Z}_2$不変量は、トポロジカル絶縁体や超伝導体のようなエキゾチックな物質を理解する上で重要であり、これらは先進的なエレクトロニクスやスピントロニクス、量子コンピューティングへの応用が期待されています[Ref.]。
+これらの不変量、例えばチャーン数や$\mathbb{Z}_2$不変量は、先進的なエレクトロニクスやスピントロニクス、量子コンピューティングへの応用が期待されているトポロジカル絶縁体や超伝導体のようなエキゾチックな物質を理解する上で重要です[Ref.]。
 本パッケージは、研究者や教育者がこれらの不変量を様々な次元や対称性クラスで計算できる、使いやすく効率的なツールセットを提供し、新たなトポロジカル相の探索と発見を促進します。
 
 
@@ -50,17 +50,18 @@ Our package provides an accessible and efficient toolset for researchers and edu
 # Statement of need
 
 固体物理学において、物質がどのような特性を持つか理解することは重要です。
-例えば、電気伝導度は物質が電気をどの程度通すかを示す重要な物理量であり、通常は、物質に弱い静電場をかけたとき物質中の大域的空間（バルク）において電子が遷移可能であるような量子状態が存在すれば有限の値を持ち、電気が流れる金属となります。
-逆に存在しなければ低い値を持ち、電気を流さない絶縁体となります。
-しかし、1980年以降、今日に至るまで大きな関心を集めている量子ホール絶縁体やトポロジカル絶縁体などを含むトポロジカル電子系は、バルクでは絶縁体でありながら、物質表面には電子の伝導状態を持つ新しい物質状態を示します[Ref.]。
-このような新奇特性から、トポロジカル電子系の候補物質の探索や特性評価がこれまで盛んに行われて来ています[@Hasan2010Colloquium;@Qi2011Topological]。
+例えば、電気伝導度は物質が電気をどの程度通すかを示す重要な物理量です。
+通常、物質に弱い静電場をかけたとき、物質中の大域的空間（バルク）において電子が遷移可能な量子状態が存在すれば、電気伝導度は有限の値を持ち、電気が流れる金属となります。
+逆に、存在しなければ電気伝導度は低い値となり、電気を流さない絶縁体となります。
+しかし、1980年以降、量子ホール絶縁体やトポロジカル絶縁体などを含むトポロジカル電子系は、バルクでは絶縁体でありながら、物質表面には電子の伝導状態を持つ新しい物質状態を示し、今日に至るまで大きな関心を集めています[@Hasan2010Colloquium;@Qi2011Topological]。
+この新奇な特性から、トポロジカル電子系の候補物質の探索や特性評価が盛んに行われてきました。
 
 
-表面の伝導状態の特徴は、物質のハミルトニアンから決定される固有量子状態が波数空間上でどのようなトポロジーを持つかによって決定されており、第一チャーン数、第二チャーン数、$\mathbb{Z}_2$不変量などのトポロジカル数がそれらを特徴付けるために用いられます[Ref.]。
[…]
