### 2024-12-04 4b7c154
update paper draft
