+++ paper/paper.md
@@ -37,17 +37,23 @@ bibliography: paper.bib
 
 # Summary
 
-`TopologicalNumbers.jl` is an open-source Julia package designed to calculate topological invariants—mathematical quantities that characterize the properties of materials—in condensed matter physics. 
-These invariants, such as the Chern number and $\mathbb{Z}_2$ invariants, are crucial for understanding exotic materials like topological insulators and superconductors which have potential applications in advanced electronics, spintronics, and quantum computing. 
-Our package provides an accessible and efficient toolset for researchers and educators to compute these invariants across various dimensions and symmetry classes, facilitating the exploration and discovery of new topological phases of matter.
+`TopologicalNumbers.jl` is an open-source Julia package designed to calculate topological invariants, which are mathematical quantities characterizing the properties of materialsin condensed matter physics. 
+These invariants, such as the Chern number and the $\mathbb{Z}_2$ invariant, are crucial for understanding exotic materials like topological insulators and superconductors, which have potential applications in advanced electronics, spintronics, and quantum computing [Ref.]. 
+This package provides researchers and educators with an easy-to-use and efficient toolset to compute these invariants across various dimensions and symmetry classes, facilitating the exploration and discovery of new topological phases of matter.
 
 `TopologicalNumbers.jl`は、凝縮系物理学における物質の性質を特徴付ける数学的量であるトポロジカル不変量を計算するための、オープンソースのJuliaパッケージです。
 これらの不変量、例えばチャーン数や$\mathbb{Z}_2$不変量は、先進的なエレクトロニクスやスピントロニクス、量子コンピューティングへの応用が期待されているトポロジカル絶縁体や超伝導体のようなエキゾチックな物質を理解する上で重要です[Ref.]。
 本パッケージは、研究者や教育者がこれらの不変量を様々な次元や対称性クラスで計算できる、使いやすく効率的なツールセットを提供し、新たなトポロジカル相の探索と発見を促進します。
 
 
-
 # Statement of need
+
+Understanding the properties of materials is essential in solid-state physics. 
+For example, electrical conductivity is a key physical quantity that indicates how well a material conducts electricity. 
+Typically, when a weak electrostatic field is applied to a material, if there exist quantum states in the bulk where electrons can transition, the material exhibits finite electrical conductivity and behaves as a metal. 
+Conversely, if such states do not exist, the electrical conductivity is low, and the material behaves as an insulator.
+Since the 1980s, topological electronic systems—including quantum Hall insulators and topological insulators—have attracted significant attention because they exhibit new states where the bulk is insulating, but the material’s surface possesses conducting electronic states [@Hasan2010Colloquium;@Qi2011Topological]. 
+Due to these novel properties, extensive research has been conducted to explore candidate materials and evaluate their characteristics.
 
 固体物理学において、物質がどのような特性を持つか理解することは重要です。
 例えば、電気伝導度は物質が電気をどの程度通すかを示す重要な物理量です。
@@ -56,6 +62,12 @@ Our package provides an accessible and efficient toolset for researchers and edu
 しかし、1980年以降、量子ホール絶縁体やトポロジカル絶縁体などを含むトポロジカル電子系は、バルクでは絶縁体でありながら、物質表面には電子の伝導状態を持つ新しい物質状態を示し、今日に至るまで大きな関心を集めています[@Hasan2010Colloquium;@Qi2011Topological]。
[…]
