### 2024-12-05 f20b5f3
update paper draft
