+++ paper/paper.md
@@ -38,7 +38,7 @@ bibliography: paper.bib
 # Summary
 
 `TopologicalNumbers.jl` is an open-source Julia package designed to calculate topological invariants, which are mathematical quantities characterizing the properties of materials in condensed matter physics. 
-These invariants, such as the Chern number and the $\mathbb{Z}_2$ number, are crucial for understanding exotic materials like topological insulators and superconductors, which have potential applications in advanced electronics, spintronics, and quantum computing [Ref.]. 
+These invariants, such as the Chern number and the $\mathbb{Z}_2$ invariant, are crucial for understanding exotic materials like topological insulators and superconductors, which have potential applications in advanced electronics, spintronics, and quantum computing [Ref.]. 
 This package provides researchers and educators with an easy-to-use and efficient toolset to compute these invariants across various dimensions and symmetry classes, facilitating the exploration and discovery of new topological phases of matter.
 
 
@@ -46,7 +46,7 @@ This package provides researchers and educators with an easy-to-use and efficien
 
 Understanding the properties of materials is essential in solid-state physics. 
 For example, electrical conductivity is a key physical quantity that indicates how well a material conducts electric current. 
-Typically, when a weak electrostatic field is applied to a material, if there exist quantum eigenstates in the bulk of material where electrons can transition, the material exhibits finite electrical conductivity and behaves as a metal. 
+Typically, when a weak electric field is applied to a material, if there exist quantum eigenstates in the bulk into which electrons can transition, the material exhibits finite electrical conductivity and behaves as a metal. 
 Conversely, if such states do not exist, the electrical conductivity is low, and the material behaves as an insulator.
 Since the 1980s, topological electronic systems—including quantum Hall insulators and topological insulators—have attracted attention because they exhibit new states where the bulk is insulating, but the material’s surface possesses conducting electronic states [@Hasan2010Colloquium;@Qi2011Topological]. 
 Due to these novel properties, extensive research has been conducted to explore candidate materials and evaluate their characteristics.
@@ -60,18 +60,18 @@ Other topological numbers similarly serve as important physical quantities chara
 
 
 To obtain the topological numbers, we often need numerical calculations, and it may require an enormous amount of computation before convergence is achieved. 
-Therefore, creating tools to easily compute these numbers will lead to advances in research on the topological phase of matter. 
-So far, several methods have been reported that suggest that some topological numbers can be computed efficiently [@Fukui2005Chern;@Fukui2007Quantum;@Mochol-Grzelak2018Efficient;@Shiozaki2023discrete]. 
-However, since each method is specialized for specific dimensions or symmetry classes, it is necessary to implement the algorithm for each problem, respectively. 
-Our project, `TopologicalNumbers.jl`, aims to provide a package that can easily and efficiently compute topological numbers comprehensively in various dimensions and symmetry classes.
+Therefore, creating tools that allow easy computation of these numbers will advance research on the topological phases of matter. 
+So far, several methods have been reported that enable efficient computation of some topological numbers [@Fukui2005Chern;@Fukui2007Quantum;@Mochol-Grzelak2018Efficient;@Shiozaki2023discrete]. 
+However, since each method is specialized for specific dimensions or symmetry classes, it is necessary to implement the algorithm separately for each problem. 
+Our project, `TopologicalNumbers.jl`, aims to provide a package that can easily and efficiently compute topological numbers across various dimensions and symmetry classes.
 
 
 This package currently includes several methods for calculating topological numbers. 
[…]
