### 2024-12-05 206970e
update paper draft
