### 2024-12-05 9c8e0ed
update paper draft
