+++ paper/paper.md
@@ -48,7 +48,8 @@ Understanding the properties of materials is essential in solid-state physics.
 For example, electrical conductivity is a key physical quantity that indicates how well a material conducts electric current. 
 Typically, when a weak electric field is applied to a material, if quantum eigenstates exist in the bulk into which electrons can transition, the material exhibits finite electrical conductivity and behaves as a metal.
 Conversely, if such states do not exist, the electrical conductivity is low, and the material behaves as an insulator.
-Since the 1980s, topological electronic systems—including quantum Hall insulators and topological insulators—have attracted attention because they exhibit new states where the bulk is insulating, but the material’s surface possesses conducting electronic states [@Hasan2010Colloquium;@Qi2011Topological]. 
+Since the 1980s, it has been revealed that some materials exhibit states where the bulk is insulating but the material’s surface has conducting electronic states [@Hasan2010Colloquium;@Qi2011Topological]. 
+These materials are known as topological electronic systems, including quantum Hall insulators and topological insulators. 
 Due to these novel properties, extensive research has been conducted to explore candidate materials and evaluate their characteristics.
 
 
@@ -56,54 +57,54 @@ The features of surface conducting states are determined by the topology of quan
 Topological numbers, such as the first Chern number, the second Chern number, and the $\mathbb{Z}_2$ invariant, are used to characterize these properties [@Thouless1982Quantized;@Kane2005Z_2]. 
 A typical example is the quantum Hall effect, where applying a weak electric field to a two-dimensional material results in a quantized finite electrical conductivity (Hall conductivity) in the direction perpendicular to the applied field [@Thouless1982Quantized]. 
 The Hall conductivity $\sigma_{xy}$ is characterized by the first Chern number $\nu \in \mathbb{Z}$ and is given by $\sigma_{xy} = \frac{e^{2}}{h} \nu$, where $e$ is the elementary charge and $h$ is Planck’s constant. 
-Other topological numbers similarly serve as important physical quantities characterizing systems, depending on their dimensions and symmetry classes [@Ryu2010Topological].
+Other topological numbers similarly serve as important physical quantities that characterize systems, depending on their dimensions and symmetry classes [@Ryu2010Topological].
 
 
-We often need numerical calculations to obtain the topological numbers, which may require enormous computation before convergence is achieved. 
-Therefore, creating tools that allow easy computation of these numbers will advance research on the topological phases of matter. 
-So far, several methods have been reported that enable efficient computation of some topological numbers [@Fukui2005Chern;@Fukui2007Quantum;@Mochol-Grzelak2018Efficient;@Shiozaki2023discrete]. 
-However, since each method is specialized for specific dimensions or symmetry classes, it is necessary to implement the algorithm separately for each problem. 
-Our project, `TopologicalNumbers.jl`, aims to provide a package that can easily and efficiently compute topological numbers across various dimensions and symmetry classes.
+Obtaining topological numbers often requires extensive numerical calculations, which may demand considerable computational effort before achieving convergence. 
+Therefore, creating tools that simplify the computation of these numbers will advance research on topological phases of matter. 
+Several methods have been developed that enable efficient computation of certain topological numbers [@Fukui2005Chern;@Fukui2007Quantum;@Mochol-Grzelak2018Efficient;@Shiozaki2023discrete]. 
+However, since each method is typically specialized for specific dimensions or symmetry classes, one must often implement algorithms separately for each problem. 
[…]
