### 2024-12-08 1b1f93c
update paper draft
