+++ paper/paper.md
@@ -37,7 +37,7 @@ bibliography: paper.bib
 
 # Summary
 
-`TopologicalNumbers.jl` is an open-source Julia package designed to calculate topological invariants, which are mathematical quantities characterizing the properties of materials in condensed matter physics. 
+`TopologicalNumbers.jl` is an open-source Julia package designed to calculate topological invariants—mathematical quantities that characterize the properties of materials in condensed matter physics. 
 These invariants, such as the Chern number and the $\mathbb{Z}_2$ invariant, are crucial for understanding exotic materials like topological insulators and superconductors, which have potential applications in advanced electronics, spintronics, and quantum computing [@Nayak2008NonAbelian;@Hasan2010Colloquium;@Qi2011Topological]. 
 This package provides researchers and educators with an easy-to-use and efficient toolset to compute these invariants across various dimensions and symmetry classes, facilitating the exploration and discovery of new topological phases of matter.
 
@@ -45,24 +45,24 @@ This package provides researchers and educators with an easy-to-use and efficien
 # Statement of need
 
 Understanding the properties of materials is essential in solid-state physics. 
-For example, electrical conductivity is a key physical quantity that indicates how well a material conducts electric current. 
+For example, electrical conductivity is a key physical quantity indicating how well a material conducts electric current. 
 Typically, when a weak electric field is applied to a material, if quantum eigenstates exist in the bulk into which electrons can transition, the material exhibits finite electrical conductivity and behaves as a metal.
 Conversely, if such states do not exist, the electrical conductivity is low, and the material behaves as an insulator.
-Since the 1980s, it has been revealed that some materials exhibit states where the bulk is insulating but the material’s surface has conducting electronic states [@Hasan2010Colloquium;@Qi2011Topological]. 
+Since the 1980s, it has been revealed that certain materials exhibit states in which the bulk is insulating while the material’s surface has conducting electronic states [@Hasan2010Colloquium;@Qi2011Topological]. 
 These materials are known as topological electronic systems, including quantum Hall insulators and topological insulators. 
-Due to these novel properties, extensive research has been conducted to explore candidate materials and evaluate their characteristics.
+Due to these novel properties, extensive research has been conducted to identify candidate materials and evaluate their characteristics.
 
 
-The features of surface conducting states are determined by the topology of quantum eigenstates in momentum space, dictated by the material’s Hamiltonian. 
+The features of surface conducting states are determined by the topology of quantum eigenstates in momentum space, determined by the material’s Hamiltonian. 
 Topological numbers, such as the first Chern number, the second Chern number, and the $\mathbb{Z}_2$ invariant, are used to characterize these properties [@Thouless1982Quantized;@Kane2005Z_2]. 
-A typical example is the quantum Hall effect, where applying a weak electric field to a two-dimensional material results in a quantized finite electrical conductivity (Hall conductivity) in the direction perpendicular to the applied field [@Thouless1982Quantized]. 
+A typical example is the quantum Hall effect, where applying a weak electric field to a two-dimensional material results in a quantized finite electrical conductivity (Hall conductivity) perpendicular to the applied field [@Thouless1982Quantized]. 
 The Hall conductivity $\sigma_{xy}$ is characterized by the first Chern number $\nu \in \mathbb{Z}$ and is given by $\sigma_{xy} = \frac{e^{2}}{h} \nu$, where $e$ is the elementary charge and $h$ is Planck’s constant. 
[…]
