### 2024-12-08 1aaefda
update paper draft
