+++ paper/paper.bib
@@ -188,7 +188,7 @@
 }
 
 @article{Yu2011Equivalent,
-  title     = {Equivalent Expression of \$\{{\textbackslash}mathbb\{\vphantom{\}\}}{{Z}}\vphantom\{\}\vphantom\{\}\_\{2\}\$ Topological Invariant for Band Insulators Using the Non-{{Abelian Berry}} Connection},
+  title     = {Equivalent Expression of $\mathbb{Z}_2$ Topological Invariant for Band Insulators Using the Non-{{Abelian Berry}} Connection},
   author    = {Yu, Rui and Qi, Xiao Liang and Bernevig, Andrei and Fang, Zhong and Dai, Xi},
   year      = {2011},
   journal   = {Phys. Rev. B},
@@ -253,43 +253,43 @@
 }
 
 @article{Haldane1988Model,
-  title = {Model for a {{Quantum Hall Effect}} without {{Landau Levels}}: {{Condensed-Matter Realization}} of the "{{Parity Anomaly}}"},
-  author = {Haldane, F. D. M.},
-  year = {1988},
-  journal = {Phys. Rev. Lett.},
-  volume = {61},
-  number = {18},
-  pages = {2015--2018},
+  title     = {Model for a {{Quantum Hall Effect}} without {{Landau Levels}}: {{Condensed-Matter Realization}} of the "{{Parity Anomaly}}"},
+  author    = {Haldane, F. D. M.},
+  year      = {1988},
+  journal   = {Phys. Rev. Lett.},
+  volume    = {61},
+  number    = {18},
+  pages     = {2015--2018},
   publisher = {American Physical Society},
-  issn = {0031-9007},
-  doi = {10.1103/PhysRevLett.61.2015},
-  urldate = {2022-06-27},
-  pmid = {10038961}
+  issn      = {0031-9007},
+  doi       = {10.1103/PhysRevLett.61.2015},
+  urldate   = {2022-06-27},
+  pmid      = {10038961}
 }
 
 @article{Su1979Solitons,
-  title = {Solitons in {{Polyacetylene}}},
-  author = {Su, W. P. and Schrieffer, J. R. and Heeger, A. J.},
-  year = {1979},
-  journal = {Phys. Rev. Lett.},
-  volume = {42},
-  number = {25},
-  pages = {1698--1701},
+  title     = {Solitons in {{Polyacetylene}}},
+  author    = {Su, W. P. and Schrieffer, J. R. and Heeger, A. J.},
+  year      = {1979},
+  journal   = {Phys. Rev. Lett.},
+  volume    = {42},
+  number    = {25},
+  pages     = {1698--1701},
   publisher = {American Physical Society},
-  doi = {10.1103/PhysRevLett.42.1698},
-  urldate = {2024-12-05}
+  doi       = {10.1103/PhysRevLett.42.1698},
+  urldate   = {2024-12-05}
 }
 
 
 @article{Nayak2008NonAbelian,
-  title = {Non-{{Abelian}} Anyons and Topological Quantum Computation},
-  author = {Nayak, Chetan and Simon, Steven H. and Stern, Ady and Freedman, Michael and Das Sarma, Sankar},
-  year = {2008},
-  journal = {Rev. Mod. Phys.},
-  volume = {80},
-  number = {3},
-  pages = {1083--1159},
+  title     = {Non-{{Abelian}} Anyons and Topological Quantum Computation},
+  author    = {Nayak, Chetan and Simon, Steven H. and Stern, Ady and Freedman, Michael and Das Sarma, Sankar},
+  year      = {2008},
+  journal   = {Rev. Mod. Phys.},
+  volume    = {80},
+  number    = {3},
[…]
