### 2024-12-09 #6944 @olexandr-konovalov
@KskAdch was just about to ask what's the status of this, and noticed activity in the repository - great to see this, hope we may see a revised version of the paper soon!
