+++ src/phaseDiagram.jl
@@ -873,49 +873,49 @@ end
 
 # Berry phase
 @doc raw"""
-Calculate the one-dimensional phase diagram for the Berry phase as a function of a parameter range.
+Calculate the one-dimensional phase diagram for the Berry phase over a specified parameter range.
 
     calcPhaseDiagram(prob::BPProblem, param_range::T1, alg::T2=BP(); parallel::T3=UseSingleThread(), plot::Bool=false, progress::Bool=false) where {T1<:AbstractVector,T2<:BerryPhaseAlgorithms,T3<:TopologicalNumbersParallel}
 
 # Description
-`calcPhaseDiagram` は、一次元系を想定した `BPProblem` に対して、指定したパラメータの範囲 (`param_range`) を走査しながら Berry 位相 を計算し、相図を得るための関数です。  
-`alg` や `parallel` オプションを切り替えることで、計算手法や並列化戦略を柔軟に変更できます。さらに、`plot` 引数を有効にすると、計算完了後に簡易的なプロットを表示することも可能です。
+`calcPhaseDiagram` calculates the Berry phase for a one-dimensional system described by `BPProblem` over a range of parameters (`param_range`), constructing a phase diagram. By changing the `alg` or `parallel` options, you can flexibly switch between different calculation methods and parallelization strategies. Moreover, setting `plot=true` generates a simple plot of the results upon completion of the calculation.
 
 # Arguments
 - `prob::BPProblem`: 
-  Berry 位相計算用の問題設定を格納した構造体です。ハミルトニアンやメッシュサイズ、ギャップの有無などのパラメータを含みます。
+  A structure that contains the problem settings for Berry phase calculations, including the Hamiltonian, mesh size, gap information, and other parameters.
 
 - `param_range::AbstractVector`: 
-  走査するパラメータのリストや範囲 (e.g. `range(-2.0, 2.0, length=50)`) を指定します。この値を用いて、内部でハミルトニアンに与えるパラメータを変化させ、Berry 位相を計算します。
+  A list or range of parameters (e.g., `range(-2.0, 2.0, length=50)`) over which to scan. The Hamiltonian is evaluated at each parameter value in order to compute the Berry phase.
 
 - `alg::BerryPhaseAlgorithms=BP()`: 
-  Berry 位相の計算アルゴリズムを指定します。既定値は `BP()` です。必要に応じて、異なるアルゴリズムを実装・指定することが可能です。
+  The algorithm used to compute the Berry phase. The default is `BP()`. You can implement and specify other algorithms as needed.
 
 - `parallel::TopologicalNumbersParallel=UseSingleThread()`: 
-  並列化の戦略を指定します。既定値はシングルスレッド (`UseSingleThread()`) です。分散環境やマルチスレッドでの実行をしたい場合は別途指定してください。（現時点でBerry Phaseの計算はマルチスレッド対応していません。）
+  The parallelization strategy. The default is single-threaded (`UseSingleThread()`). If you wish to run in a distributed environment, you can specify it accordingly. (Note that, at present, the Berry phase calculation does not support multithreading.)
 
 - `plot::Bool=false`: 
-  `true` に設定すると、計算完了後に簡易的なプロットが表示されます。可視化によって位相遷移を直感的に確認したい場合に有用です。
+  If set to `true`, a simple plot of the calculated results is shown after the computation finishes. This is helpful for visually confirming topological phase transitions.
 
 - `progress::Bool=false`: 
-  `true` に設定すると、ループ計算時に進捗状況を表示します。計算規模が大きい場合の目安として活用できます。
[…]
