### 2025-01-20 71ccb8a
Update version to 1.7.6 in Project.toml
