+++ Project.toml
@@ -1,7 +1,7 @@
 name = "TopologicalNumbers"
 uuid = "82b459e2-751c-486c-8b84-4a85e9d7ad51"
 authors = ["Keisuke Adachi <18s2002x@gmail.com>", "Minoru Kanega <phys_chibaraki@yahoo.co.jp>"]
-version = "1.7.5"
+version = "1.7.6"
 
 [deps]
 Accessors = "7d9f7c33-5ae7-4f3b-8dc6-eff91059b697"
