### 2025-01-20 8d67efa
Update compatibility versions in Project.toml
