+++ Project.toml
@@ -16,12 +16,12 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
 Accessors = "0.1"
-Distributed = "^1.10"
+Distributed = "^1.6"
 LaTeXStrings = "1"
-LinearAlgebra = "^1.10"
+LinearAlgebra = "^1.6"
 Parameters = "0.12"
 ProgressBars = "1.5"
 PythonPlot = "1.0"
-SparseArrays = "^1.10"
+SparseArrays = "^1.6"
 StaticArrays = "1"
-julia = "^1.10"
+julia = "^1.6"
