### 2025-01-21 df795e2
Change PythonPlot version to 1.0.3 in Project.toml for AutoMerge CI
