+++ Project.toml
@@ -21,7 +21,7 @@ LaTeXStrings = "1"
 LinearAlgebra = "^1.6"
 Parameters = "0.12"
 ProgressBars = "1.5"
-PythonPlot = "1.0"
+PythonPlot = "1.0.3"
 SparseArrays = "^1.6"
 StaticArrays = "1"
 julia = "^1.6"
