### 2025-01-21 de82095
Add CondaPkg dependency and update PythonPlot version in Project.toml
