+++ test/Project.toml
@@ -15,4 +15,4 @@ JuliaFormatter = "1.0"
 LinearAlgebra = "^1.6"
 MPI = "0.20"
 PythonCall = "0.9"
-PythonPlot = "1.0.3"
+PythonPlot = "1.0"
