+++ CondaPkg.toml
@@ -1,0 +1,3 @@
+[deps]
+matplotlib = ""
+python = ">=3.11,<3.12"
