+++ Project.toml
@@ -5,6 +5,7 @@ version = "1.7.6"
 
 [deps]
 Accessors = "7d9f7c33-5ae7-4f3b-8dc6-eff91059b697"
+CondaPkg = "992eb4ea-22a4-4c89-a5bb-47a3300528ab"
 Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
@@ -16,12 +17,13 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
 Accessors = "0.1"
+CondaPkg = "0.2"
 Distributed = "^1.6"
 LaTeXStrings = "1"
 LinearAlgebra = "^1.6"
 Parameters = "0.12"
 ProgressBars = "1.5"
-PythonPlot = "1.0.3"
+PythonPlot = "1.0"
 SparseArrays = "^1.6"
 StaticArrays = "1"
 julia = "^1.6"
