### 2025-01-21 8a0efb1
Enable Agg backend for PythonPlot in AutoMerge CI environment
