+++ src/packages.jl
@@ -4,6 +4,11 @@ using LaTeXStrings
 using LinearAlgebra
 using Parameters: @unpack
 using ProgressBars
-using PythonPlot
 using SparseArrays
 using StaticArrays
+
+# For AutoMerge CI
+if haskey(ENV, "GITHUB_ACTIONS") && haskey(ENV, "AUTOMERGE_GITHUB_TOKEN")
+    ENV["MPLBACKEND"] = "Agg"
+end
+using PythonPlot
