### 2025-01-21 612ccf3
Remove Python version constraint in CondaPkg.toml and add MPLBACKEND setting in packages.jl
