+++ CondaPkg.toml
@@ -1,3 +1,3 @@
 [deps]
 matplotlib = ""
-python = ">=3.11,<3.12"
+python = ""
