+++ v1.7.6/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"2D/flux/#Two-dimensional-square-lattice-model-with-flux","page":"Square Lattice w/ Flux","title":"Two-dimensional square lattice model with flux","text":"","category":"section"},{"location":"2D/flux/","page":"Square Lattice w/ Flux","title":"Square Lattice w/ Flux","text":"A two-dimensional example is presented here:","category":"page"},{"location":"2D/flux/","page":"Square Lattice w/ Flux","title":"Square Lattice w/ Flux","text":"julia> function H₀(k, p)\n           k1, k2 = k\n           Hsize, ν = p\n           t = 1\n\n           Hmat = zeros(ComplexF64, Hsize, Hsize)\n\n           ϕ = 2π * ν / Hsize\n\n           for i in 1:Hsize\n               Hmat[i, i] = -2t * cos(k2 - i * ϕ)\n           end\n\n           for i in 1:Hsize-1\n               Hmat[i, i+1] = -t\n               Hmat[i+1, i] = -t\n           end\n\n           Hmat[1, Hsize] = -t * exp(-im * k1)\n           Hmat[Hsize, 1] = -t * exp(im * k1)\n\n           Hmat\n       end","category":"page"},{"location":"2D/flux/","page":"Square Lattice w/ Flux","title":"Square Lattice w/ Flux","text":"You can also use our preset Hamiltonian function Flux2d to define the same Hamiltonian matrix as follows:","category":"page"},{"location":"2D/flux/","page":"Square Lattice w/ Flux","title":"Square Lattice w/ Flux","text":"julia> H₀(k, p) = Flux2d(k, p)","category":"page"},{"location":"2D/flux/","page":"Square Lattice w/ Flux","title":"Square Lattice w/ Flux","text":"To calculate the dispersion, run:","category":"page"},{"location":"2D/flux/","page":"Square Lattice w/ Flux","title":"Square Lattice w/ Flux","text":"julia> H(k) = H₀(k, (6, 1))\njulia> showBand(H; value=false, disp=true)","category":"page"},{"location":"2D/flux/","page":"Square Lattice w/ Flux","title":"Square Lattice w/ Flux","text":"(Image: Dispersion of 2D square lattice with flux model)","category":"page"},{"location":"2D/flux/","page":"Square Lattice w/ Flux","title":"Square Lattice w/ Flux","text":"Then we can compute the Chern numbers using FCProblem:","category":"page"},{"location":"2D/flux/","page":"Square Lattice w/ Flux","title":"Square Lattice w/ Flux","text":"julia> prob = FCProblem(H);\njulia> sol = solve(prob)","category":"page"},{"location":"2D/flux/","page":"Square Lattice w/ Flux","title":"Square Lattice w/ Flux","text":"The output is:","category":"page"},{"location":"2D/flux/","page":"Square Lattice w/ Flux","title":"Square Lattice w/ Flux","text":"FCSolution{Vector{Int64}, Int64}([1, 1, -2, -2, 1, 1], 0)","category":"page"},{"location":"2D/flux/","page":"Square Lattice w/ Flux","title":"Square Lattice w/ Flux","text":"The first argument TopologicalNumber in the named tuple is an vector that stores the first Chern number for each band.  The vector is arranged in order of bands, starting from the one with the lowest energy. The second argument Total stores the total of the first Chern numbers for each band. Total is a quantity that should always return zero.","category":"page"},{"location":"2D/flux/","pag
[…]
