### 2025-01-21 #6944 @phjmsycc
@olexandr-konovalov @lf28 @michiexile

I hope you are doing well. Thank you very much for dedicating your time and effort to review our manuscript.

We submitted our revised version on December 11, which was about 40 days ago. In addition, nearly a year has passed since our initial submission (around January 27 of last year), and we would greatly appreciate any update on the current status of the review or the anticipated timeline moving forward.

We truly value the substantial amount of time and effort you have already invested in this process. If there is any further information, clarification, or additional work we can provide, please do not hesitate to let us know. Thank you again for your assistance, and we look forward to your response!
