### 2025-02-15 #6944 @phjmsycc
Please find below the version and DOI information:
- Version number: v1.7.6
- DOI: https://doi.org/10.5281/zenodo.14710087
