### 2025-02-26 #6944 @olexandr-konovalov
@phjmsycc thanks - hope to look next week, when we will have no classes.
