### 2025-04-11 #6944 @olexandr-konovalov
@phjmsycc I've proofread the paper - could you please merge https://github.com/KskAdch/TopologicalNumbers.jl/pull/80 ?
