### 2025-04-22 #6944 @kyleniemeyer
Hi @phjmsycc @KskAdch could you please add an explicit References section header (`# References`) at the end of the paper? That is the only issue I see before accepting.
