+++ paper/paper.md
@@ -113,4 +113,4 @@ The authors are grateful to Takahiro Fukui for fruitful discussions.
 M. K. was supported by JST, the establishment of university fellowships towards the creation of science technology innovation (Grant No. JPMJFS2107), and by JST SPRING (Grant No. JPMJSP2109).
 
 
-#References
+# References
